+++ apps/api/src/app/endpoints/watchlist/watchlist.controller.ts
@@ -1,12 +1,13 @@
 import { HasPermission } from '@ghostfolio/api/decorators/has-permission.decorator';
 import { Impersonation } from '@ghostfolio/api/decorators/impersonation.decorator';
+import { RequiresScope } from '@ghostfolio/api/decorators/requires-scope.decorator';
 import { HasPermissionGuard } from '@ghostfolio/api/guards/has-permission.guard';
-import { ImpersonationGuard } from '@ghostfolio/api/guards/impersonation.guard';
 import { TransformDataSourceInRequestInterceptor } from '@ghostfolio/api/interceptors/transform-data-source-in-request/transform-data-source-in-request.interceptor';
 import { TransformDataSourceInResponseInterceptor } from '@ghostfolio/api/interceptors/transform-data-source-in-response/transform-data-source-in-response.interceptor';
 import { CreateWatchlistItemDto } from '@ghostfolio/common/dtos';
 import { WatchlistResponse } from '@ghostfolio/common/interfaces';
 import { permissions } from '@ghostfolio/common/permissions';
+import { scopes } from '@ghostfolio/common/scopes';
 import {
   ImpersonationContext,
   RequestWithUser
@@ -82,7 +83,7 @@ export class WatchlistController {
 
   @Get()
   @HasPermission(permissions.readWatchlist)
-  @UseGuards(AuthGuard('jwt'), HasPermissionGuard, ImpersonationGuard)
+  @RequiresScope(scopes.watchlistRead)
   @UseInterceptors(TransformDataSourceInResponseInterceptor)
   public async getWatchlistItems(
     @Impersonation() { userId }: ImpersonationContext