+++ apps/client/src/app/components/access-table/access-table.component.ts
@@ -1,6 +1,7 @@
 import { ConfirmationDialogType } from '@ghostfolio/common/enums';
 import { Access, User } from '@ghostfolio/common/interfaces';
 import { publicRoutes } from '@ghostfolio/common/routes/routes';
+import { hasScope, scopes } from '@ghostfolio/common/scopes';
 import { NotificationService } from '@ghostfolio/ui/notifications';
 
 import { Clipboard, ClipboardModule } from '@angular/cdk/clipboard';
@@ -98,6 +99,10 @@ export class GfAccessTableComponent {
     const languageCode = this.user().settings.language;
 
     return `${this.baseUrl}/${languageCode}/${publicRoutes.public.path}/${aId}`;
+  }
+
+  protected hasScopeToReadValues({ scopes: scopesOfAccess }: Access) {
+    return hasScope(scopesOfAccess, scopes.portfolioReadValues);
   }
 
   protected onCopyUrlToClipboard(aId: string) {