+++ libs/common/src/lib/permissions.ts
@@ -1,6 +1,6 @@
 import { UserWithSettings } from '@ghostfolio/common/types';
 
-import { Access, Role } from '@prisma/client';
+import { Role } from '@prisma/client';
 
 export const permissions = {
   accessAdminControl: 'accessAdminControl',
@@ -196,24 +196,6 @@ export function hasPermission(
   aPermission: string
 ) {
   return aPermissions.includes(aPermission);
-}
-
-export function hasReadRestrictedAccessPermission({
-  accesses = [],
-  impersonationId
-}: {
-  accesses?: Pick<Access, 'id' | 'permissions'>[];
-  impersonationId: string | null;
-}) {
-  if (!impersonationId) {
-    return false;
-  }
-
-  const access = accesses.find(({ id }) => {
-    return id === impersonationId;
-  });
-
-  return access?.permissions?.includes('READ_RESTRICTED') ?? true;
 }
 
 export function hasRole(aUser: UserWithSettings, aRole: Role) {