+++ libs/common/src/lib/types/user-with-settings.type.ts
@@ -1,11 +1,10 @@
 import { SubscriptionType } from '@ghostfolio/common/enums';
 import { SubscriptionOffer, UserSettings } from '@ghostfolio/common/interfaces';
 
-import { Access, Account, Settings, User } from '@prisma/client';
+import { Account, Settings, User } from '@prisma/client';
 
 // TODO: Compare with User interface
 export type UserWithSettings = User & {
-  accessesGet: Access[];
   accounts: Account[];
   activityCount: number;
   dataProviderGhostfolioDailyRequests: number;